# docs/ebay-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "eBay Scraper API",
    "description": "Extract sold listings, active listings, and price data from eBay",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "Health check endpoint.",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/sold": {
      "post": {
        "summary": "Get Sold Listings",
        "description": "Get sold/completed listings from eBay.\n\nReturns actual sale prices (what items sold for, not asking prices).\nGreat for price research, market analysis, and arbitrage.",
        "operationId": "get_sold_listings_sold_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SoldListingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoldListingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Sold Listings Get",
        "description": "GET endpoint for sold listings (convenience for simple requests).",
        "operationId": "get_sold_listings_get_sold_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Search query",
              "title": "Query"
            },
            "description": "Search query"
          },
          {
            "name": "min_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum price",
              "title": "Min Price"
            },
            "description": "Minimum price"
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum price",
              "title": "Max Price"
            },
            "description": "Maximum price"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Max results",
              "default": 50,
              "title": "Limit"
            },
            "description": "Max results"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "eBay category ID",
              "title": "Category"
            },
            "description": "eBay category ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoldListingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/active": {
      "post": {
        "summary": "Get Active Listings",
        "description": "Get active (non-sold) listings from eBay.\n\nReturns current listings including auctions and buy-it-now.",
        "operationId": "get_active_listings_active_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActiveListingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveListingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Active Listings Get",
        "description": "GET endpoint for active listings (convenience for simple requests).",
        "operationId": "get_active_listings_get_active_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Search query",
              "title": "Query"
            },
            "description": "Search query"
          },
          {
            "name": "min_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum price",
              "title": "Min Price"
            },
            "description": "Minimum price"
          },
          {
            "name": "max_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum price",
              "title": "Max Price"
            },
            "description": "Maximum price"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Max results",
              "default": 50,
              "title": "Limit"
            },
            "description": "Max results"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "eBay category ID",
              "title": "Category"
            },
            "description": "eBay category ID"
          },
          {
            "name": "listing_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "'auction' or 'buy_it_now'",
              "title": "Listing Type"
            },
            "description": "'auction' or 'buy_it_now'"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveListingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActiveListingsRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Search query"
          },
          "min_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Price",
            "description": "Minimum price filter"
          },
          "max_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Price",
            "description": "Maximum price filter"
          },
          "limit": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "Maximum results to return",
            "default": 50
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "description": "eBay category ID"
          },
          "listing_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Type",
            "description": "'auction' or 'buy_it_now'"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "ActiveListingsRequest"
      },
      "ActiveListingsResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "listings": {
            "items": {
              "$ref": "#/components/schemas/ListingItem"
            },
            "type": "array",
            "title": "Listings"
          },
          "scraped_at": {
            "type": "string",
            "title": "Scraped At"
          }
        },
        "type": "object",
        "required": [
          "query",
          "count",
          "listings",
          "scraped_at"
        ],
        "title": "ActiveListingsResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ListingItem": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "ebay_item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ebay Item Id"
          },
          "sold_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sold Date"
          },
          "bids": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bids"
          },
          "time_left": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Left"
          }
        },
        "type": "object",
        "required": [
          "title",
          "price",
          "url"
        ],
        "title": "ListingItem"
      },
      "PriceStats": {
        "properties": {
          "median_price": {
            "type": "number",
            "title": "Median Price"
          },
          "avg_price": {
            "type": "number",
            "title": "Avg Price"
          },
          "min_price": {
            "type": "number",
            "title": "Min Price"
          },
          "max_price": {
            "type": "number",
            "title": "Max Price"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "median_price",
          "avg_price",
          "min_price",
          "max_price",
          "count"
        ],
        "title": "PriceStats"
      },
      "SoldListingsRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Search query",
            "example": "PSA 10 Charizard"
          },
          "min_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Price",
            "description": "Minimum price filter"
          },
          "max_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Price",
            "description": "Maximum price filter"
          },
          "limit": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Limit",
            "description": "Maximum results to return",
            "default": 50
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "description": "eBay category ID"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SoldListingsRequest"
      },
      "SoldListingsResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "listings": {
            "items": {
              "$ref": "#/components/schemas/ListingItem"
            },
            "type": "array",
            "title": "Listings"
          },
          "stats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PriceStats"
              },
              {
                "type": "null"
              }
            ]
          },
          "scraped_at": {
            "type": "string",
            "title": "Scraped At"
          }
        },
        "type": "object",
        "required": [
          "query",
          "count",
          "listings",
          "scraped_at"
        ],
        "title": "SoldListingsResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RapidAPI-Proxy-Secret"
      }
    }
  }
}